services/marketplace-service/requirements.txt
Django==4.2.9
djangorestframework==3.14.0
PyJWT==2.8.0
psycopg2-binary==2.9.9
redis==5.0.1
django-redis==5.4.0
python-decouple==3.8
celery==5.3.6
celery[redis]==5.3.6
django-celery-beat==2.5.0
drf-spectacular==0.27.1
django-cors-headers==4.3.1
gunicorn==21.2.0
httpx==0.26.0
channels==4.0.0
channels-redis==4.1.0
